Rotate the pointer/needle to North
I've looked all around the internet for code that worked to make a 3D compass. I do not know much about Quaternions, I've messed around with a simulation and found that editing w and y would yield the result I wanted if done right. The rest was trial and error :